TCA-55 Moving the tests and other domain files
