+++ app/src/main/java/ar/scacchipa/twittercloneapp/data/repository/IAuthorizationRepository.kt
@@ -1,7 +1,5 @@
 package ar.scacchipa.twittercloneapp.data.repository
 
-import ar.scacchipa.twittercloneapp.data.ResponseDomain
-
 interface IAuthorizationRepository {
     suspend fun requestAccessToken(
         transitoryToken: String,
