+++ app/src/main/java/ar/scacchipa/twittercloneapp/di/AppModule.kt
@@ -1,19 +1,19 @@
 package ar.scacchipa.twittercloneapp.di
 
 import ar.scacchipa.twittercloneapp.data.IMapper
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenData
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenDataMapper
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenDomain
+import ar.scacchipa.twittercloneapp.data.datasource.UserAccessTokenData
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDataMapper
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDomain
 import ar.scacchipa.twittercloneapp.data.datasource.IAuthDataSource
 import ar.scacchipa.twittercloneapp.domain.AuthorizationUseCase
 import ar.scacchipa.twittercloneapp.domain.ConsumableAuthUseCase
 import ar.scacchipa.twittercloneapp.domain.SplashTimerUseCase
-import ar.scacchipa.twittercloneapp.presentation.AuthWebDialog.FragmentAuthWebDialog
+import ar.scacchipa.twittercloneapp.presentation.authwebdialog.FragmentAuthWebDialog
 import ar.scacchipa.twittercloneapp.presentation.login.FragmentLogin
 import ar.scacchipa.twittercloneapp.presentation.splash.FragmentSplash
 import ar.scacchipa.twittercloneapp.data.repository.AuthorizationRepository
 import ar.scacchipa.twittercloneapp.data.repository.IAuthorizationRepository
-import ar.scacchipa.twittercloneapp.presentation.AuthWebDialog.AuthWebDialogViewModel
+import ar.scacchipa.twittercloneapp.presentation.authwebdialog.AuthWebDialogViewModel
 import ar.scacchipa.twittercloneapp.presentation.login.LoginViewModel
 import ar.scacchipa.twittercloneapp.presentation.splash.SplashViewModel
 import org.koin.androidx.fragment.dsl.fragment
