+++ app/src/main/java/ar/scacchipa/twittercloneapp/presentation/authwebdialog/AuthWebDialogViewModel.kt
@@ -1,4 +1,4 @@
-package ar.scacchipa.twittercloneapp.presentation.AuthWebDialog
+package ar.scacchipa.twittercloneapp.presentation.authwebdialog
 
 import android.webkit.WebResourceError
 import android.webkit.WebViewClient
@@ -6,7 +6,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import ar.scacchipa.twittercloneapp.data.ResponseDomain
+import ar.scacchipa.twittercloneapp.data.repository.ResponseDomain
 import ar.scacchipa.twittercloneapp.domain.AuthorizationUseCase
 import ar.scacchipa.twittercloneapp.domain.ConsumableAuthUseCase
 import ar.scacchipa.twittercloneapp.utils.Constants
