+++ app/src/main/res/layout/fragment_auth_web_dialog_layout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/secondary_background"
-    tools:context=".presentation.AuthWebDialog.FragmentAuthWebDialog">
+    tools:context=".presentation.authwebdialog.FragmentAuthWebDialog">
 
     <WebView
         android:id="@+id/login_webview"
