--- app/src/test/java/ar/scacchipa/twittercloneapp/component/.gitkeep
@@ -1,1 +1,0 @@
-//To track the empty directory
