+++ app/src/test/java/ar/scacchipa/twittercloneapp/data/UserAccessTokenDataMapperTest.kt
@@ -1,5 +1,8 @@
 package ar.scacchipa.twittercloneapp.data
 
+import ar.scacchipa.twittercloneapp.data.datasource.UserAccessTokenData
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDataMapper
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDomain
 import org.junit.Assert
 import org.junit.Test
