+++ app/src/test/java/ar/scacchipa/twittercloneapp/data/repository/AuthorizationRepositoryTest.kt
@@ -1,9 +1,9 @@
 package ar.scacchipa.twittercloneapp.repository
 
-import ar.scacchipa.twittercloneapp.data.ResponseDomain
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenData
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenDataMapper
-import ar.scacchipa.twittercloneapp.data.UserAccessTokenDomain
+import ar.scacchipa.twittercloneapp.data.repository.ResponseDomain
+import ar.scacchipa.twittercloneapp.data.datasource.UserAccessTokenData
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDataMapper
+import ar.scacchipa.twittercloneapp.data.repository.UserAccessTokenDomain
 import ar.scacchipa.twittercloneapp.data.repository.AuthorizationRepository
 import ar.scacchipa.twittercloneapp.data.repository.IAuthorizationRepository
 import ar.scacchipa.twittercloneapp.data.datasource.IAuthDataSource
@@ -22,7 +22,8 @@ class AuthorizationRepositoryTest {
 
     private var subject: IAuthorizationRepository = AuthorizationRepository(
         MockAuthDataSource(),
-        UserAccessTokenDataMapper())
+        UserAccessTokenDataMapper()
+    )
 
     @Test
     fun repoShouldReturnASuccessToken() = runTest {
